refactored some place model classes
