+++ src/main/java/pl/warkoczewski/foodspot/fetcher/PlaceFetcher.java
@@ -1,10 +1,10 @@
 package pl.warkoczewski.foodspot.fetcher;
 
 import pl.warkoczewski.foodspot.model.PlaceQuery;
-import pl.warkoczewski.foodspot.model.place.Result;
+import pl.warkoczewski.foodspot.model.place.Place;
 
 import java.util.List;
 
 public interface PlaceFetcher {
-    List<Result> getResults(PlaceQuery placeQuery);
+    List<Place> fetchPlaces(PlaceQuery placeQuery);
 }
