+++ src/main/java/pl/warkoczewski/foodspot/model/place/Record.java
@@ -1,0 +1,84 @@
+
+package pl.warkoczewski.foodspot.model.place;
+
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import com.fasterxml.jackson.annotation.JsonAnyGetter;
+import com.fasterxml.jackson.annotation.JsonAnySetter;
+import com.fasterxml.jackson.annotation.JsonIgnore;
+import com.fasterxml.jackson.annotation.JsonInclude;
+import com.fasterxml.jackson.annotation.JsonProperty;
+import com.fasterxml.jackson.annotation.JsonPropertyOrder;
+
+@JsonInclude(JsonInclude.Include.NON_NULL)
+@JsonPropertyOrder({
+    "html_attributions",
+    "next_page_token",
+    "results",
+    "status"
+})
+public class Record {
+
+    @JsonProperty("html_attributions")
+    private List<Object> htmlAttributions = null;
+    @JsonProperty("next_page_token")
+    private String nextPageToken;
+    @JsonProperty("results")
+    private List<Place> places = null;
+    @JsonProperty("status")
+    private String status;
+    @JsonIgnore
+    private Map<String, Object> additionalProperties = new HashMap<String, Object>();
+
+    @JsonProperty("html_attributions")
+    public List<Object> getHtmlAttributions() {
+        return htmlAttributions;
+    }
+
+    @JsonProperty("html_attributions")
+    public void setHtmlAttributions(List<Object> htmlAttributions) {
+        this.htmlAttributions = htmlAttributions;
+    }
+
+    @JsonProperty("next_page_token")
+    public String getNextPageToken() {
+        return nextPageToken;
+    }
+
+    @JsonProperty("next_page_token")
+    public void setNextPageToken(String nextPageToken) {
+        this.nextPageToken = nextPageToken;
+    }
+
+    @JsonProperty("results")
+    public List<Place> getPlaces() {
+        return places;
+    }
+
+    @JsonProperty("results")
+    public void setPlaces(List<Place> places) {
+        this.places = places;
+    }
+
+    @JsonProperty("status")
+    public String getStatus() {
+        return status;
+    }
+
+    @JsonProperty("status")
+    public void setStatus(String status) {
+        this.status = status;
+    }
+
+    @JsonAnyGetter
+    public Map<String, Object> getAdditionalProperties() {
+        return this.additionalProperties;
+    }
+
+    @JsonAnySetter
+    public void setAdditionalProperty(String name, Object value) {
+        this.additionalProperties.put(name, value);
+    }
+
+}
