--- src/main/java/pl/warkoczewski/foodspot/model/place/Result.java
@@ -1,188 +1,0 @@
-
-package pl.warkoczewski.foodspot.model.place;
-
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-import com.fasterxml.jackson.annotation.JsonAnyGetter;
-import com.fasterxml.jackson.annotation.JsonAnySetter;
-import com.fasterxml.jackson.annotation.JsonIgnore;
-import com.fasterxml.jackson.annotation.JsonInclude;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import com.fasterxml.jackson.annotation.JsonPropertyOrder;
-
-@JsonInclude(JsonInclude.Include.NON_NULL)
-@JsonPropertyOrder({
-    "geometry",
-    "icon",
-    "id",
-    "name",
-    "opening_hours",
-    "photos",
-    "place_id",
-    "price_level",
-    "rating",
-    "reference",
-    "types",
-    "vicinity"
-})
-public class Result {
-
-    @JsonProperty("geometry")
-    private Geometry geometry;
-    @JsonProperty("icon")
-    private String icon;
-    @JsonProperty("id")
-    private String id;
-    @JsonProperty("name")
-    private String name;
-    @JsonProperty("opening_hours")
-    private OpeningHours openingHours;
-    @JsonProperty("photos")
-    private List<Photo> photos = null;
-    @JsonProperty("place_id")
-    private String placeId;
-    @JsonProperty("price_level")
-    private Integer priceLevel;
-    @JsonProperty("rating")
-    private Double rating;
-    @JsonProperty("reference")
-    private String reference;
-    @JsonProperty("types")
-    private List<String> types = null;
-    @JsonProperty("vicinity")
-    private String vicinity;
-    @JsonIgnore
-    private Map<String, Object> additionalProperties = new HashMap<String, Object>();
-
-    @JsonProperty("geometry")
-    public Geometry getGeometry() {
-        return geometry;
-    }
-
-    @JsonProperty("geometry")
-    public void setGeometry(Geometry geometry) {
-        this.geometry = geometry;
-    }
-
-    @JsonProperty("icon")
-    public String getIcon() {
-        return icon;
-    }
-
-    @JsonProperty("icon")
-    public void setIcon(String icon) {
-        this.icon = icon;
-    }
-
-    @JsonProperty("id")
-    public String getId() {
-        return id;
-    }
-
-    @JsonProperty("id")
-    public void setId(String id) {
-        this.id = id;
-    }
-
-    @JsonProperty("name")
-    public String getName() {
-        return name;
-    }
-
-    @JsonProperty("name")
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    @JsonProperty("opening_hours")
-    public OpeningHours getOpeningHours() {
-        return openingHours;
-    }
-
-    @JsonProperty("opening_hours")
-    public void setOpeningHours(OpeningHours openingHours) {
-        this.openingHours = openingHours;
-    }
-
-    @JsonProperty("photos")
-    public List<Photo> getPhotos() {
-        return photos;
-    }
-
-    @JsonProperty("photos")
-    public void setPhotos(List<Photo> photos) {
-        this.photos = photos;
-    }
-
-    @JsonProperty("place_id")
-    public String getPlaceId() {
-        return placeId;
-    }
-
-    @JsonProperty("place_id")
-    public void setPlaceId(String placeId) {
-        this.placeId = placeId;
-    }
-
-    @JsonProperty("price_level")
-    public Integer getPriceLevel() {
-        return priceLevel;
-    }
-
-    @JsonProperty("price_level")
-    public void setPriceLevel(Integer priceLevel) {
-        this.priceLevel = priceLevel;
-    }
-
-    @JsonProperty("rating")
-    public Double getRating() {
-        return rating;
-    }
-
-    @JsonProperty("rating")
-    public void setRating(Double rating) {
-        this.rating = rating;
-    }
-
-    @JsonProperty("reference")
-    public String getReference() {
-        return reference;
-    }
-
-    @JsonProperty("reference")
-    public void setReference(String reference) {
-        this.reference = reference;
-    }
-
-    @JsonProperty("types")
-    public List<String> getTypes() {
-        return types;
-    }
-
-    @JsonProperty("types")
-    public void setTypes(List<String> types) {
-        this.types = types;
-    }
-
-    @JsonProperty("vicinity")
-    public String getVicinity() {
-        return vicinity;
-    }
-
-    @JsonProperty("vicinity")
-    public void setVicinity(String vicinity) {
-        this.vicinity = vicinity;
-    }
-
-    @JsonAnyGetter
-    public Map<String, Object> getAdditionalProperties() {
-        return this.additionalProperties;
-    }
-
-    @JsonAnySetter
-    public void setAdditionalProperty(String name, Object value) {
-        this.additionalProperties.put(name, value);
-    }
-
-}
