+++ src/main/java/pl/warkoczewski/foodspot/service/PlaceServiceImpl.java
@@ -5,7 +5,6 @@ import org.springframework.stereotype.Service;
 import pl.warkoczewski.foodspot.dto.DisplayPlaceDTO;
 import pl.warkoczewski.foodspot.dto.PlaceQueryDTO;
 import pl.warkoczewski.foodspot.fetcher.PlaceFetcherImpl;
-import pl.warkoczewski.foodspot.model.PLACE_TYPE;
 import pl.warkoczewski.foodspot.model.PlaceQuery;
 
 import java.util.List;
@@ -20,16 +19,9 @@ public class PlaceServiceImpl implements PlaceService {
         this.placeFetcher = placeFetcher;
         this.modelMapper = modelMapper;
     }
-
-    public List<DisplayPlaceDTO> getResulPlaces(){
-        PlaceQueryDTO placeQueryDTO = new PlaceQueryDTO("52.4064", "16.9252"
-                , 15000l, PLACE_TYPE.RESTAURANT, "Italian");
-        return getResults(placeQueryDTO);
-    }
-
     @Override
-    public List<DisplayPlaceDTO> getResults(PlaceQueryDTO placeQueryDTO) {
-        return placeFetcher.getResults(modelMapper.map(placeQueryDTO, PlaceQuery.class))
+    public List<DisplayPlaceDTO> getPlaces(PlaceQueryDTO placeQueryDTO) {
+        return placeFetcher.fetchPlaces(modelMapper.map(placeQueryDTO, PlaceQuery.class))
                 .stream()
                 .map(result -> new DisplayPlaceDTO(result.getName())).collect(Collectors.toList());
     }
