+++ src/main/resources/templates/place/places.html
@@ -49,7 +49,7 @@
 </div><br><br><br><br>
 <div class="container">
     <div th:if="${#lists.isEmpty(places)}">
-        <h2>Sorry, no results</h2>
+        <h2>Sorry, no places</h2>
     </div>
 
     <div th:if="${not #lists.isEmpty(places)}">
