+++ src/test/java/pl/warkoczewski/foodspot/fetcher/RecordFetcherImplTest.java
@@ -4,14 +4,17 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
+import pl.warkoczewski.foodspot.dto.PlaceQueryDTO;
+import pl.warkoczewski.foodspot.model.PLACE_TYPE;
 
 import static org.junit.jupiter.api.Assertions.*;
 @SpringBootTest
 @ExtendWith(SpringExtension.class)
-class PlaceFetcherImplTest {
+class RecordFetcherImplTest {
     @Test
     void getApiURL_Should_Return_Given_Url() {
         //given
+
         String expectedUrl = "https://maps.googleapis.com/maps/api/place/nearbysearch/json?" +
                 "location=52.4064,16.9252&radius=150000&type=restaurant&" +
                 "keyword=Italian&key=AIzaSyAzjbfgQXOHrxMwqh7JS253V--zbXJtqmU";
@@ -22,6 +25,8 @@ class PlaceFetcherImplTest {
 
     @Test
     void getResults() {
+        PlaceQueryDTO placeQueryDTO = new PlaceQueryDTO("52.4064", "16.9252"
+                , 15000l, PLACE_TYPE.RESTAURANT, "Italian");
     }
